Use new JS Solr Utils library
The code has been updated to use the new JS Solr Utils library that
reduces the code and makes the partial be cleaner.

This commit has the following changes:

Update code related to Kmaps jQuery Plugins
Updated the code to reflect changes on the Kmaps jQuery plugins
Relations Tree and Collapsible List

Update KmapsPopup to use solr-utils library
Updated the code from jquery.kmaps-popup.js to now use the solr-utils.js
library this to keep all the solr queries in one place and simplify the
code. So the code in the _related plartial was updated to reflect this
change.

Update code to on related for summary
Cleaned code for new update on solr-utils library