Using the new popovers on the flyout search results
The new version of the popovers can be used now on the search result,
check kmaps_engine for the changes, now it allows to display the related
assets to the left of the element so it shows properly on the search
results.

Update code to take advantage of popup language support

Kmaps-popup now takes into account language, so this update takes
advantage of that.

Updated flyout and related features trees

A new version of the tree is being provided to standardize the use of
trees, all of them will be handled by kmapsRelationsTree so the code has
been updated to reflect this change.

Fixed locale for bo, dz, en, zh and updated version in preparation for merge to master