Adds Shapes, Altitudes, and centroid to Solr documents
When indexing a feature in Solr, we now include the GeoData in GeoJSON
format and Also the Altitudes.

Both pieces of information (Altitudes and GeoData) are stored as
subdocuments. Solr schema needs to be updated to support field types for
Geometries using the dynamic field `_grptgeom`.

Adds the ability to obtain centroids for shapes by feature

Updated version in preparation for merge

Adds GeoJSON representation as expected by ArcGIS

ArcGIS API uses GeoJSON this way:

https://developers.arcgis.com/javascript/latest/api-reference/esri-layers-GeoJSONLayer.html

There are a few problems with what it expects and how we generate
GeoJSON.

It is expecting a `FeatureCollection`, so this works:

{ "type":"FeatureCollection",
  "features": [
    {
      "type":"Feature",
      "geometry": {
        "type":"Point",
        "coordinates":[91.1359083974058,29.6430462172591]
      }
    }
  ]
}

We were only generating the Features:

{
  "type":"Feature",
  "geometry":{
    "type":"Point",
    "coordinates":[91.1359083974058,29.6430462172591]
  }
}

Another problem when we are using `MultiPlolygon` is that it expects the
geometries to comply with the Right Hand Rule, which is ambiguos. We are
using PostGIS to generate the GeoJSON so If I want to rotate it
ClockWise it doesn't work:

ST_ForceRHR should be the transformation which is an alias for
ST_ForcePolygonCW

The way it works is if We force a Counter Clockwise

ST_ForcePolygonCCW

We also need to add the `FeaturCollection` part to the GeoJSON, which is
what ArcGIS expects. So we are building a query similar to this:

SELECT json_build_object(
     'type', 'FeatureCollection',
     'features', json_agg(
         json_build_object(
             'type',       'Feature',
             'id',         gid, -- the GeoJson spec includes an 'id' field, but it is optional, replace {id} with your id field
             'geometry',   ST_AsGeoJSON(ST_ForcePolygonCCW(geometry))::json,
             'properties', json_build_object(
                 -- list of fields
                 'area', area,
                 'altitude', altitude
             )
         )
     )
 ) FROM shapes where gid = 12752; --Lhasa in this case

So we are calling json_build_object from the query.

Add type to Geometries for Solr documents